style: Refactor code
* Split business components
* Rename Activity Header Component
* Add some Styling details to some elements like buttons and text
