+++ src/views/People/ActivityInfo/Activity/Activity.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PastActivities from "./PastActivities";
-import ActivityDetailHeader from "./ActivityDetailHeader";
+import ActivityDetailHeader from "./ActivityHeader";
 import styled from "styled-components";
-import { getColor } from "../../../_starter/theme/theme";
+import { getColor } from "../../../../_starter/theme/theme";
 
 
 const ActivitiesList = styled.ul`
