+++ src/views/People/ActivityInfo/Activity/ItemDate.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import dateFormat from "dateformat";
 import styled from "styled-components";
-import { getColor } from "../../../_starter/theme/theme";
+import { getColor } from "../../../../_starter/theme/theme";
 
 const DateContainer = styled.p`
   display: block;
