+++ src/views/People/ActivityInfo/Reminders/Reminders.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Reminders() {
+  return (
+    <div></div>
+  );
+}
+
+export default Reminders;
