+++ src/views/People/ActivityInfo/Traking/Traking.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Traking() {
+  return (
+    <div></div>
+  );
+}
+
+export default Traking;
